minor: further clean up code
